src/requirements.txt
motor==3.7.1
pymongo==4.14.0
python-dotenv==1.1.1
quart==0.20.0
shortuuid==1.0.13
uvicorn==0.35.0
pytz==2025.2